profile the Mortal-path rollout properly and fix the actual bottleneck (episode payload), not the one I guessed
Measured, in order, after two wrong guesses:
  * stubbing the encoder to a constant changes throughput by 1.9% -- feature
    extraction is NOT the bottleneck, so no C/CUDA rewrite is warranted;
  * bisecting depth vs input width: 3->40 blocks costs -36%, 21->934 planes
    costs -70%, and the GPU forward is identical for both widths (20.3 vs
    20.6 ms), so it is not compute either;
  * the host-side batch path copies the same 32 MB three times (stack 4.7 +
    copy 1.7 + pageable H2D 3.4 = 9.9 ms, vs 1.2 ms from pinned memory);
  * but the dominant cost is the episode payload: 11.1 MB per game at 934
    planes vs 0.3 MB at 21, i.e. ~23 GB of trajectories built, pickled and
    shipped per 2048-game iteration. That is what OOM-killed exp41 arm B.

Fix: ship episode planes as float16 (every plane is an indicator, probability
or rescaled ratio in [0,1], and the update forward runs under bf16 autocast
anyway) -- 11.1 -> 5.66 MB per game. Encoder micro-optimisations kept (1.13x:
dict tile tables replacing ~340 str.replace per encode, dora set hoisted out
of a 4x redundant recompute, ~200 scattered numpy writes batched into two
vectorised scatters), pinned by a golden-checksum test so any future change
that perturbs a byte fails loudly.

+++ scripts/train_dnn_ppo.py
@@ -322,7 +322,8 @@ def main():
         games += len(results)
 
         cat = lambda k: np.concatenate([e[k] for e in episodes])
-        planes = torch.from_numpy(cat("planes")).to(dev)
+        # episodes ship planes as float16 (see _package_game); widen once here
+        planes = torch.from_numpy(cat("planes")).to(dev).float()
         scal = torch.from_numpy(cat("scalars")).to(dev)
         mask = torch.from_numpy(cat("mask")).to(dev)
         acts = torch.from_numpy(cat("actions")).to(dev)
